Worked on the tagging filesystem and did some cleanup.
Tagging filesystem should now be able to fully handle the files directory with meta data and all.
I also started work on the tags directory and made the root file object loadable.
Removed the unused VirtualMeta from base since it really didn't follow the new meta data structure and rather than make it conform it was easier to delete it given that it was unused.
Did some cleanup of the vfswebdavhandler and removed the get_prop_* methods that were used by the unbranched version of the code.

Known problems: The tags filesystem has not been under any kind of test other than basic syntax validation so it is probably not working at all.